requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alpha_vantage==3.0.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.6.12
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
frozendict==2.4.6
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.3
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
yfinance==0.2.54
